# docs/swagger.yaml
basePath: /v1
definitions:
  domain.Column:
    properties:
      id:
        readOnly: true
        type: string
      name:
        type: string
      position:
        type: integer
      project_id:
        type: string
      status:
        type: string
    required:
    - name
    - project_id
    - status
    type: object
...
  domain.Project:
    properties:
      description:
        type: string
      id:
        readOnly: true
        type: string
      name:
        type: string
    required:
    - description
    - name
    type: object
  domain.Task:
    properties:
      column_id:
        type: string
      description:
        type: string
      id:
        readOnly: true
        type: string
      name:
        type: string
      position:
        type: integer
    required:
    - column_id
    - description
    - name
    type: object
  web.HTTPError:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
info:
  contact: {}
  title: Task Tracker API
  version: "1.0"
paths:
  /columns:
    get:
      description: get all columns
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Column'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.HTTPError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/web.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.HTTPError'
      summary: Get all columns
      tags:
      - columns
    post:
      consumes:
      - application/json
      description: add by json column
      parameters:
      - description: Add column
        in: body
        name: project
        required: true
        schema:
          $ref: '#/definitions/domain.Column'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Column'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.HTTPError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/web.HTTPError'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/web.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.HTTPError'
      summary: Add a column
      tags:
      - columns
  /columns/{id}:
    delete:
      description: Delete by column ID
      parameters:
      - description: column ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: it's ok
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.HTTPError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/web.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.HTTPError'
      summary: Delete a column
      tags:
      - columns
    get:
      description: get column by id
      parameters:
      - description: column ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Column'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.HTTPError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/web.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.HTTPError'
      summary: Show a column
      tags:
      - columns
...
  /projects:
    get:
      description: get all projects
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Project'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.HTTPError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/web.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.HTTPError'
      summary: Get all projects
      tags:
      - projects
    post:
      consumes:
      - application/json
      description: add by json project
      parameters:
      - description: Add project
        in: body
        name: project
        required: true
        schema:
          $ref: '#/definitions/domain.Project'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Project'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.HTTPError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/web.HTTPError'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/web.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.HTTPError'
      summary: Add a project
      tags:
      - projects
  /projects/{id}:
    delete:
      description: Delete by project ID
      parameters:
      - description: project ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: it's ok
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.HTTPError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/web.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.HTTPError'
      summary: Delete a project
      tags:
      - projects
    get:
      description: get project by id
      parameters:
      - description: project ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Project'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.HTTPError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/web.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.HTTPError'
      summary: Show a project
      tags:
      - projects
  /projects/{id}/columns:
    get:
      description: get columns by project id
      parameters:
      - description: project ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Column'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.HTTPError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/web.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.HTTPError'
      summary: Get columns by project id
      tags:
      - projects
  /projects/{id}/tasks:
    get:
      description: get tasks by project id
      parameters:
      - description: project ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Task'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.HTTPError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/web.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.HTTPError'
      summary: Get tasks by project id
      tags:
      - projects
  /tasks:
    get:
      description: get all tasks
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Task'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.HTTPError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/web.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.HTTPError'
      summary: Get all tasks
      tags:
      - tasks
    post:
      consumes:
      - application/json
      description: add by json task
      parameters:
      - description: Add task
        in: body
        name: project
        required: true
        schema:
          $ref: '#/definitions/domain.Task'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Task'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.HTTPError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/web.HTTPError'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/web.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.HTTPError'
      summary: Add a task
      tags:
      - tasks
  /tasks/{id}:
    delete:
      description: Delete by task ID
      parameters:
      - description: task ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: it's ok
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.HTTPError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/web.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.HTTPError'
      summary: Delete a task
      tags:
      - tasks
    get:
      description: get task by id
      parameters:
      - description: task ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Task'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.HTTPError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/web.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.HTTPError'
      summary: Show a task
      tags:
      - tasks
...
swagger: "2.0"